requirements.txt
agno>=0.4.0
httpx>=0.27.0
python-dotenv>=1.0.1
lancedb>=0.5.7
sqlalchemy>=2.0.25
psycopg[binary]>=3.2.1
fastapi>=0.110.0
uvicorn[standard]>=0.23.0
pandas>=2.0.0
pyarrow>=15.0.0
openai>=1.0.0
